requirements.txt
azure-ai-vision-imageanalysis
Flask
python-dotenv
azure-storage-blob
azure-identity